feat: final refact and check
