+++ README.md
@@ -1,7 +1,7 @@
 # Проект: Место
 
 ### Описание
-Учебный проект о местах России. Задачи проекта: отработка навыков адаптивной верстки страниц, верстки форм и модальных окон, добавление функциональности (открытие/закрытие модальных окон, добавление/удаление карточек, валидация и отправка форм), разделения кода на модули, настройки сборки Webpack, взаимодействия с сервером.
+Учебный проект о местах России. Исходные задачи проекта: отработка навыков адаптивной верстки страниц, верстки форм и модальных окон, добавление функциональности (открытие/закрытие модальных окон, добавление/удаление карточек, валидация и отправка форм), разделения кода на модули, настройки сборки Webpack, взаимодействия с сервером. Проект переписан в логике ООП.
 
 ### Технологии
 * HTML5
