+++ src/components/Card.js
@@ -1,4 +1,5 @@
 export default class Card {
+
   constructor({ data, handleCardClick, handleLikeClick, handleDeleteClick }, selector, userId) {
     this._name = data.name;
     this._link = data.link;
@@ -12,6 +13,7 @@ export default class Card {
     this._handleLikeClick = handleLikeClick;
   }
 
+  // метод для получения элемента
   _getElement() {
     const cardElement = document
         .querySelector(this._selector)
@@ -22,6 +24,7 @@ export default class Card {
     return cardElement;
   }
 
+  // метод для создания карточки
   generate() {
     this._element = this._getElement();
     this._element.querySelector('.articles__title').textContent = this._name;
@@ -57,6 +60,7 @@ export default class Card {
     return this._element;
   }
 
+  // метод для удаления карточки
   delete() {
     this._element.remove();
   }
@@ -76,80 +80,5 @@ export default class Card {
     this._element.querySelector('.articles__likes-count').textContent = cardNewData.likes.length;
     this._isLiked = false;
   }
+
 }
-
-
-
-// import { openPopup } from './modal.js';
-// import { popupPic, popupPicPicture, popupPicCaption, articleTemplate } from './index.js';
-// import { deleteCard, likeCard, unlikeCard } from './Api.js';
-
-
-// // Функция создания карточки
-
-// export function createCard(card, user) {
-//     //const articleElement = articleTemplate.querySelector('.articles__item').cloneNode(true);
-//     //const articleTitle = articleElement.querySelector('.articles__title');
-//     //const articlePhoto = articleElement.querySelector('.articles__photo');
-//     //const articleLikes = articleElement.querySelector('.articles__likes-count');
-//     //articleTitle.textContent = card.name;
-//     //articlePhoto.src = card.link;
-//     //articlePhoto.alt = card.name;
-//     //articleLikes.textContent = card.likes.length;
-//     articleElement.setAttribute("data-id", card._id);
-//     articleElement.setAttribute("data-owner-id", card.owner._id);
-
-//     // функционал лайка
-//     // const iconLike = articleElement.querySelector('.articles__like');
-//     // if (card.likes.some((obj) => {
-//     //   return obj._id === user;
-//     // })) {
-//     //   iconLike.classList.add('articles__like_active');
-//     // }
-//     iconLike.addEventListener('click', function() {
-//       if (!iconLike.classList.contains('articles__like_active')) {
-//         likeCard(articleElement)
-//         .then((cardNewData) => {
-//           iconLike.classList.add('articles__like_active');
-//           articleLikes.textContent = cardNewData.likes.length;
-//         })
-//         .catch((err) => {
-//           console.log(err);
-//         });
-//       } else {
-//         unlikeCard(articleElement)
-//         .then((cardNewData) => {
-//           iconLike.classList.remove('articles__like_active');
-//           articleLikes.textContent = cardNewData.likes.length;
-//         })
-//         .catch((err) => {
-//           console.log(err);
-//         });
-//       }
-//     });
-
-//     // функционал удаления
-//     const iconDelete = articleElement.querySelector('.articles__delete');
-//     if (card.owner._id !== user) {
-//       iconDelete.remove();
-//     } else {
-//       iconDelete.addEventListener('click', function() {
-//         deleteCard(articleElement)
-//         .then(() => {
-//           articleElement.remove();
-//         })
-//         .catch((err) => {
-//           console.log(err);
-//         });
-//       });
-//     }
-
-//     articlePhoto.addEventListener('click', function() {
-//       openPopup(popupPic);
-//       popupPicPicture.src = articlePhoto.src;
-//       popupPicPicture.alt = articleTitle.textContent;
-//       popupPicCaption.textContent = articleTitle.textContent;
-//     });
-    
-//     return articleElement;
-// };
