--- src/components/utils.js
@@ -1,11 +1,0 @@
-// лоадер на кнопки
-
-// перенесено в PopupWithForm
-
-// export function renderLoading(isLoading, button) {
-//   if(isLoading) {
-//     button.innerText = 'Сохранение...';
-//   } else {
-//     setTimeout(function() {button.innerText = button.dataset.text}, 300);
-//   }
-// }
